refactor: Move Result into its own file; change to oauthURI; add APIType.community to eventually replace profile
