+++ BattleNetAPI/Model/WebService/Network.swift
@@ -14,64 +14,6 @@ enum HTTPMethod: String {
 }
 
 
-enum Result<T> {
-    case success(T)
-    case failure(HTTPError)
-    
-    /// Returns `true` if the result is a success, `false` otherwise.
-    public var isSuccess: Bool {
-        switch self {
-        case .success:
-            return true
-        case .failure:
-            return false
-        }
-    }
-    
-    
-    
-    // MARK: - Init
-    
-    init(value: T?, error: HTTPError?) {
-        switch (value, error) {
-        case (let v?, _):
-            self = .success(v)
-        case (nil, let e?):
-            self = .failure(e)
-        case (nil, nil):
-            let error = HTTPError(type: .httpError, code: 1, description: "Invalid input: value and error were both nil.")
-            self = .failure(error)
-        }
-    }
-}
-
-
-
-extension Result where T == Data {
-    /**
-     Decode a successful Result to a class specified in your completion. Will return an HTTPError.jsonParsingError if decoding fails, or the original error is the Result is a failure.
-     
-     - parameter customDecode: Implement a customDecode method to return the desired Decodable object if it's nested in the data or needs to be configured
-     - parameter completion: Called when decoding is complete, or when errors are encountered
-    */
-    func decode<U: Decodable>(customDecode: ((_ data: Data) throws -> U)? = nil, completion: @escaping (_ result: Result<U>) -> Void) {
-        switch self {
-        case .success(let data):
-            do {
-                let decoded = try customDecode?(data) ?? U.decode(from: data)
-                completion(.success(decoded))
-            }
-            catch {
-                completion(.failure(HTTPError(type: .jsonParsingError)))
-            }
-        case .failure(let error):
-            completion(.failure(error))
-        }
-    }
-}
-
-
-
 class Network {
     static let shared = Network()
     private init() { }
@@ -104,7 +46,7 @@ class Network {
         request.setValue("application/json", forHTTPHeaderField: "Content-Type")
         request.setValue("application/json", forHTTPHeaderField: "Accept")
         
-        if apiType == .profile,
+        if apiType == .profile || apiType == .community,
             let userAccessToken = userAccessToken {
             request.setValue(String(format: "Bearer %@", userAccessToken), forHTTPHeaderField: "Authorization")
         }
